pgsql_py_prj1/requirements.txt
anyio==3.6.2
brotlipy==0.7.0
certifi==2021.10.8
cffi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/croot-2p6q7e2r/cffi_1642701115063/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.3
conda==4.11.0
conda-content-trust @ file:///tmp/build/80754af9/conda-content-trust_1617045594566/work
conda-package-handling @ file:///opt/concourse/worker/volumes/live/8fb3e065-760b-4a9d-4cd9-aca7fc8baf53/volume/conda-package-handling_1618262145611/work
cryptography @ file:///opt/concourse/worker/volumes/live/a9ca20dd-e1fe-4ed3-754e-0db64082a5d2/volume/cryptography_1639414576087/work
fastapi==0.89.1
h11==0.14.0
httptools==0.5.0
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
Jinja2==3.1.2
MarkupSafe==2.1.2
psycopg2-binary==2.9.5
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==1.10.4
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1635333100036/work
PySocks @ file:///opt/concourse/worker/volumes/live/112288ac-9cb0-4e73-768b-13baf4ca6419/volume/pysocks_1605305820043/work
python-dotenv==0.21.1
python-multipart==0.0.5
PyYAML==6.0
requests @ file:///opt/conda/conda-bld/requests_1641824580448/work
ruamel-yaml-conda @ file:///opt/concourse/worker/volumes/live/e81cf0fe-611a-498e-6e69-a7320057c1ac/volume/ruamel_yaml_1616016689696/work
six @ file:///tmp/build/80754af9/six_1623709665295/work
sniffio==1.3.0
starlette==0.22.0
tqdm @ file:///tmp/build/80754af9/tqdm_1635330843403/work
typing_extensions==4.4.0
urllib3==1.26.7
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
